requirements.txt
requests
pydantic
pydantic-settings
pandas
numpy
pyyaml
